requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
python-multipart==0.0.20
bcrypt==4.2.1
PyJWT==2.10.1
aioboto3==13.3.0
jinja2==3.1.5
itsdangerous==2.2.0
starlette==0.41.3
